Duke class: add method abstraction
